WIP abstraction layer for rendering
